Explicitly annotate all benchmark classes w/ State
Currently, not all benchmarks could be executed on native targets
because only a few of them are explicitly annotated with @State.
The majority of benchmarks inherit that annotation from a base class.
It works fine w/ JMH, but doesn't work for native targets.

The change will make all benchmarks executable on native targets.