Organize Buffer's segments as a regular list
Previously, Buffer's segments were organized into a circular list.
That allowed storing only a single reference to buffer's head,
and also facilitated insertion/removal of new list nodes.
The downside of a circular list is that one has
to always compare a current node with a head when
iterating over segments.
That complicates the implementation of a public API
for segments iterations.

See https://github.com/Kotlin/kotlinx-io/issues/135#issuecomment-2125014081
for details on segment iteration API.