[Unsafe API 3/5] Implement Unsafe API to iterate over segments and access its data (#336)
This is a second portion of unsafe API described in #135.
This particular part aimed to facilitate the implementation of various Buffer extensions.

The implemented API was initially described in "2. Buffer iteration API" and "3. Byte-granular read/write API" subsections of #135 (comment)

This PR concludes the changes required to implement the API from #135, but our journey won't end here! The next patch in the queue is rewriting all existing extensions using the new API and then encapsulating all segments' internals.